uncolor() fix - when arrayref is passed as the first argument.
uncolor stores all passed arguments in its lexical @data array.

If the first element of @data is arraref, script will dereference it
and push its data to the end of @data array, but will keep the arrayref
as the first argument, which leads to bad output.

This fix makes sure that this arrayref is not kept in the @data.